+++ lib/utils.ts
@@ -1,6 +1,7 @@
 import { type ClassValue, clsx } from "clsx";
 import { twMerge } from "tailwind-merge";
 import { account } from "@/appwrite/client";
+import { format, parse } from "date-fns";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
@@ -26,4 +27,24 @@ export const isAdmin = async () => {
   } catch (e) {
     return false;
   }
+};
+
+export const parseDate = (date: string) => {
+  let parsedDate = parse(date, "d-M-yyyy", new Date()).toString();
+
+  parsedDate === "Invalid Date" &&
+    (parsedDate = parse(date, "d MMM yyyy", new Date()).toString());
+
+  parsedDate === "Invalid Date" &&
+    (parsedDate = parse(
+      date,
+      "yyyy-MM-dd'T'HH:mm:ss.SSSxxx",
+      new Date(),
+    ).toString());
+
+  parsedDate = format(parsedDate, "yyyy-MM-dd'T'HH:mm:ss.SSSxxx").toString();
+
+  (parsedDate = format(parsedDate, "do MMMM yyyy")).toString();
+
+  return parsedDate;
 };