+++ changelog.md
@@ -1,7 +1,11 @@
+<!-- @format -->
+
 # Changelogs
+
 _Please refer to Git Commit History for previous Logs_
 
 ### 3 May 2024
+
 - Added Default Banner
 - Modified Create Event Page
   - Added field for Banners
@@ -12,10 +16,12 @@ _Please refer to Git Commit History for previous Logs_
 - Added Gallery Page (Placeholder)
 
 ### 4 May 2024
+
 - Added moving gradient background for home page
 - Changed Hero Section Image
 
 ### 10 May 2024
+
 - Updated Date picker for Create Event Page
 - Added new default open graph image
 - Tweaks
@@ -26,11 +32,19 @@ _Please refer to Git Commit History for previous Logs_
 - Added Dialogs
 
 ### 12 May 2024
+
 - Login Page
   - Fixed Account not found bug in Login Page
   - Added callback URL for Login Page
 - Improved Matcher for middleware
 
-#### TODO
-- Convert Login and Register Page to Server Components and add metadata.
-- Make changes to date formatting in events.
+#### ~~TODO~~
+
+~~- Convert Login and Register Page to Server Components and add metadata.~~
+~~- Make changes to date formatting in events.~~
+
+### 13 May 2024
+
+- Added Metadata to Login and Register
+- Made date formatting consistent across all events
+- Code Refactored